make intel mpi wrappers variables
the prior setup for mpi with toolkit intel was too restrictive
in that it forced the use of the wrappers mpiifort/mpiicc/mpiicpc.

while that is certainly needed for some if not all of the intel mpi
versions, that creates problems if we are using mpich or openmpi
versions which use the intel compilers as backend.

now in the makefiles the variables

INTEL_MPI_WRAPPER_FC
INTEL_MPI_WRAPPER_CC
INTEL_MPI_WRAPPER_CPP

are defined and can then be set by the user to override the intel
mpi wrappers