remove dependency on static libs to other libs, update gcc version check
remove reference to static libraries (*.a) from different components
when creating the static libraries for DSVP, INTERVIRT and INTRAVIRT.

before this change and with xcode 15, the fact that libraries for
some modules (e.g.  libtalsh.a, libddss.a etc) were being added to
the libraries of other modules confused the linker when building
the test programs.

i also don't see why one would add .a files into other .a files, if
at the end the point is to link against each of the .a files explicitly.

if this change does not cause side effect on other, non-mac osx  installations
it can  be applied elsewhere.

also slightly changed the conditions to check for gnu compiler versions
that require the -fallow-argument-mismatch argument